+++ CHANGELOG.md
@@ -8,6 +8,29 @@
 
 ### Changed
 
+## [v0.0.6] - 2015-03-25
+
+### Fixed
+
+- removed dos2unix dependency for the update command
+
+### Added
+
+- doctor command added
+- docker-check-version command added
+- cci-latest accepts branch as parameter, needed for PR testing
+- export fn command in DEBUG mode
+- export env command in DEBUG mode
+- doctor: add instruction about setting DOCKER_XXX env vars in Profile
+- info() function added to print green text to STDOUT
+
+### Removed
+
+### Changed
+
+- HOME env var is also inherited (boot2docker version failed)
+- release process fully automatized
+
 ## [v0.0.5] - 2015-03-23
 
 - `update` command works without dos2unix
@@ -17,7 +40,7 @@
 ### Fixed
 
 - debug function fixed
-- DEBUG and TRACE env vars are inherited
+- DEBUG, TRACE and CBD_DEFAULT_PROFILE env vars are inherited
 
 ### Added
 
@@ -34,7 +57,7 @@
 
 - Tool specific library renamed from cloudbreak.bash to deployer.bash
 
-## Unreleased
+## [v0.0.6] - 2015-03-25
 
 ## [v0.0.3] - 2015-03-19